Add window size configuration in preferences
This changes the current behaviour by allowing the user to change the
window size if the "remember window size" setting is not activated.

If it is, then the user won't be able to change the size (as it will be
updated each time the window size changes), but the preferences will
still display the current size; which is cool.